+++ rollup.config.js
@@ -1,13 +1,15 @@
 import resolve from 'rollup-plugin-node-resolve';
 import commonjs from 'rollup-plugin-commonjs';
 import { terser } from 'rollup-plugin-terser';
+import sizes from 'rollup-plugin-sizes';
 
 module.exports = {
   input: 'src/index.js',
   plugins: [
     resolve(),
     commonjs(),
- //   terser(),
+    terser(),
+    sizes(),
   ],
   output: {
     file: 'lib/index.js',